+++ bofire/outlier_detection/api.py
@@ -1,5 +1,0 @@
-from bofire.outlier_detection.mapper import map
-from bofire.outlier_detection.outlier_detection import (
-    IterativeTrimming,
-    OutlierDetection,
-)