+++ bofire/strategies/predictives/qparego.py
@@ -6,7 +6,6 @@ from botorch.acquisition.acquisition import AcquisitionFunction
 from botorch.acquisition.objective import GenericMCObjective
 from botorch.utils.multi_objective.scalarization import get_chebyshev_scalarization
 from botorch.utils.sampling import sample_simplex
-from pydantic import PositiveInt
 from typing_extensions import Self
 
 from bofire.data_models.acquisition_functions.acquisition_function import (
@@ -22,7 +21,6 @@ from bofire.data_models.objectives.api import (
     MinimizeObjective,
     Objective,
 )
-from bofire.data_models.outlier_detection.outlier_detections import OutlierDetections
 from bofire.data_models.strategies.api import QparegoStrategy as DataModel
 from bofire.data_models.strategies.convergence_criteria.api import (
     AnyConvergenceCriterion,
@@ -161,9 +159,6 @@ class QparegoStrategy(BotorchStrategy):
         acquisition_function: qEI | qLogEI | qLogNEI | qNEI | None = None,
         acquisition_optimizer: AnyAcqfOptimizer | None = None,
         surrogate_specs: BotorchSurrogates | None = None,
-        outlier_detection_specs: OutlierDetections | None = None,
-        min_experiments_before_outlier_check: PositiveInt | None = None,
-        frequency_check: PositiveInt | None = None,
         frequency_hyperopt: int | None = None,
         folds: int | None = None,
         seed: int | None = None,
@@ -185,9 +180,6 @@ class QparegoStrategy(BotorchStrategy):
             acquisition_function: The acquisition function to use.
             acquisition_optimizer: The optimizer for the acquisition function.
             surrogate_specs: Specifications for the surrogate model.
-            outlier_detection_specs: Specifications for outlier detection.
-            min_experiments_before_outlier_check: Minimum number of experiments before checking for outliers.
-            frequency_check: Frequency of outlier checks.
             frequency_hyperopt: Frequency of hyperparameter optimization.
             folds: Number of folds for cross-validation for hyperparameter optimization.
             seed: Random seed for reproducibility.